refactor: Columns 객체 역할 확대
단순히 columns 와 pkColumns 를 반환하는 역할만 있었는데,
종민님의 [피드백](https://github.com/next-step/jpa-entity-manager/pull/122#discussion_r1509890130) 을 반영해
곳곳에 퍼져있던 List<Column> 을 이용하는 역할을
Columns 객체에게 위임하였다.
