+++ src/main/java/persistence/sql/dml/Update.java
@@ -4,7 +4,6 @@ import persistence.sql.mapping.Columns;
 import persistence.sql.mapping.Table;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class Update {
 
@@ -17,9 +16,7 @@ public class Update {
     public Update(final Table table) {
         this.table = table;
         this.columns = new Columns(table.getColumns());
-        final List<Where> wheres = this.columns.getPkColumns().stream()
-                .map(column -> new Where(column, column.getValue(), LogicalOperator.AND, new ComparisonOperator(ComparisonOperator.Comparisons.EQ)))
-                .collect(Collectors.toList());
+        final List<Where> wheres = this.columns.generatePkColumnWheres();
         this.whereClause = new Wheres(wheres);
     }
 
@@ -32,9 +29,7 @@ public class Update {
     }
 
     public String columnSetClause() {
-        return columns.getColumns().stream()
-                .map(column -> column.getName() + " = " + column.getValue().getValueClause())
-                .collect(Collectors.joining(", "));
+        return columns.columnSetClause();
     }
 
 }
